requirements.txt
streamlit
pandas==2.2.1
matplotlib==3.8.4
seaborn==0.13.2
plotly.express==0.4.1
scikit-learn==1.4.2
numpy==1.26.4